+++ spec/acceptance/tags_spec.rb
@@ -169,15 +169,8 @@ resource 'Tags' do
                              {
                                  expected_json_path: 'data/0/type_of_tag',
                                  data_item_count: 1,
-                                 regex_match: [
-                                     /"taggings":\[\{"id":\d+,"audio_event_id":\d+,/,
-                                     /"created_at":"[^"]+"/,
-                                     /"updated_at":"[^"]+"/,
-                                     /"creator_id":\d+/,
-                                     /"updater_id":null/
-                                 ],
-                                 response_body_content: "\"taggings\":[{\"",
-                                 invalid_content: "\"taggings\":[\""
+                                 response_body_content: "general",
+                                 invalid_content: "\"taggings\":[{\""
                               })
   end